requirements.txt
ag2==0.9.10
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
autogen==0.9.10
autogen-agentchat==0.7.5
autogen-core==0.7.5
autogen-ext==0.7.5
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.2
idna==3.10
importlib_metadata==8.7.0
jiter==0.11.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langsmith==0.4.33
mcp==1.16.0
openai==2.0.0
openinference-instrumentation==0.1.39
openinference-instrumentation-autogen==0.1.9
openinference-instrumentation-openai==0.1.37
openinference-semantic-conventions==0.1.23
opentelemetry-api==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
packaging==25.0
pillow==11.3.0
protobuf==5.29.5
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
termcolor==3.1.0
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
wrapt==1.17.3
zipp==3.23.0
zstandard==0.25.0
